docs: migrated `variable-outputs`
